add resource spec interface to share between TaskConfig and ServiceConfig
