+++ src/dependency-manager/src/component-config/builder.ts
@@ -4,7 +4,7 @@ import { plainToClass } from 'class-transformer';
 import fs from 'fs-extra';
 import yaml from 'js-yaml';
 import path from 'path';
-import { DeploySpec } from '../service-config/base';
+import { DeploySpec } from '../common/base';
 import { Dictionary } from '../utils/dictionary';
 import { flattenValidationErrorsWithLineNumbers, ValidationErrors } from '../utils/errors';
 import { insertFileDataFromRefs } from '../utils/files';
