+++ src/dependency-manager/src/index.ts
@@ -1,6 +1,7 @@
 import DependencyManager from './manager';
 
 export default DependencyManager;
+export * from './common/base';
 export * from './component-config/base';
 export * from './component-config/builder';
 export * from './environment-config/base';
