+++ src/dependency-manager/src/manager.ts
@@ -112,7 +112,7 @@ export default abstract class DependencyManager {
       graph.addNode(node);
     }
 
-    const ref_map: Dictionary<string> = {};
+    const service_ref_map: Dictionary<string> = {};
     // Load component services
     for (const [service_name, service_config] of Object.entries(prefixed_component.getServices())) {
       const node_config = service_config.copy();
@@ -123,10 +123,10 @@ export default abstract class DependencyManager {
       });
       graph.addNode(node);
 
-      ref_map[service_name] = node.ref;
+      service_ref_map[service_name] = node.ref;
     }
 
-    // const task_map: Dictionary<string> = {};
+    const task_ref_map: Dictionary<string> = {};
     // Load component tasks
     for (const [task_name, task_config] of Object.entries(prefixed_component.getTasks())) {
       const node_config = task_config.copy();
@@ -137,7 +137,7 @@ export default abstract class DependencyManager {
       });
       graph.addNode(node);
 
-      // ref_map[task_name] = node.ref; TODO:84:we're not going to worry about adding the tasks to the ref_map yet, we'll need to do this though to get them working with edges
+      task_ref_map[task_name] = node.ref;
     }
 
     // Load component dependencies
@@ -152,10 +152,9 @@ export default abstract class DependencyManager {
       await this.loadComponent(graph, dep_component, component_map);
     }
 
-    // TODO:84:add edges to and from tasks
-    // Add edges to services inside component
+    // Add edges FROM services to other services and dependencies
     for (const [service_name, service_config] of Object.entries(component.getServices())) {
-      const from = ref_map[service_name];
+      const from = service_ref_map[service_name];
       const from_node = graph.getNodeByRef(from);
       if (from_node.is_external) {
         continue;
@@ -170,7 +169,7 @@ export default abstract class DependencyManager {
       let matches;
       while ((matches = services_regex.exec(service_string)) != null) {
         const [_, service_name, interface_name] = matches;
-        const to = ref_map[service_name];
+        const to = service_ref_map[service_name];
         if (to === from) continue;
         if (!service_edge_map[to]) service_edge_map[to] = {};
         service_edge_map[to]['service'] = interface_name;
@@ -201,6 +200,54 @@ export default abstract class DependencyManager {
       }
     }
 
+    // Add edges FROM tasks to other services and dependencies
+    for (const [task_name, task_config] of Object.entries(component.getTasks())) {
+      const from = task_ref_map[task_name];
+      const from_node = graph.getNodeByRef(from);
+      if (from_node.is_external) {
+        continue;
+      }
+
+      let task_string = serialize(task_config);
+      task_string = replaceBrackets(task_string);
+
+      // Add edges between services inside the component
+      const services_regex = new RegExp(`\\\${{\\s*services\\.(${Slugs.ArchitectSlugRegexNoMaxLength})?\\.interfaces\\.(${Slugs.ArchitectSlugRegexNoMaxLength})?\\.`, 'g');
+      const task_edge_map: Dictionary<Dictionary<string>> = {};
+      let matches;
+      while ((matches = services_regex.exec(task_string)) != null) {
+        const [_, service_name, interface_name] = matches;
+        const to = service_ref_map[service_name];
+        if (to === from) continue;
+        if (!task_edge_map[to]) task_edge_map[to] = {};
+        task_edge_map[to]['service'] = interface_name;
+      }
+      for (const [to, interfaces_map] of Object.entries(task_edge_map)) {
+        const edge = new ServiceEdge(from, to, interfaces_map);
+        graph.addEdge(edge);
+      }
+
+      // Add edges between services and dependencies inside the component
+      const dependencies_regex = new RegExp(`\\\${{\\s*dependencies\\.(${ComponentSlugUtils.RegexNoMaxLength})?\\.interfaces\\.(${Slugs.ArchitectSlugRegexNoMaxLength})?\\.`, 'g');
+      const dep_edge_map: Dictionary<Dictionary<string>> = {};
+      while ((matches = dependencies_regex.exec(task_string)) != null) {
+        const [_, dep_name, interface_name] = matches;
+        const dep_tag = component.getDependencies()[dep_name];
+
+        const dep_component = component_map[`${dep_name}:${dep_tag}`];
+        const to = dep_component.getInterfacesRef();
+        if (!graph.nodes_map.has(to)) continue;
+
+        if (!dep_edge_map[to]) dep_edge_map[to] = {};
+        dep_edge_map[to]['service'] = interface_name;
+      }
+
+      for (const [to, interfaces_map] of Object.entries(dep_edge_map)) {
+        const edge = new ServiceEdge(from, to, interfaces_map);
+        graph.addEdge(edge);
+      }
+    }
+
     // Add edges between services and the component's interfaces node
     const service_edge_map: Dictionary<Dictionary<string>> = {};
     for (const [component_interface_name, component_interface] of Object.entries(component.getInterfaces())) {
@@ -211,7 +258,7 @@ export default abstract class DependencyManager {
       if (!matches) continue;
 
       const [_, service_name, interface_name] = matches;
-      const to = ref_map[service_name];
+      const to = service_ref_map[service_name];
       if (!service_edge_map[to]) service_edge_map[to] = {};
       service_edge_map[to][component_interface_name] = interface_name;
     }
@@ -375,7 +422,6 @@ export default abstract class DependencyManager {
         context[normalized_ref].architect = architect_context;
       }
 
-      // TODO:84:add support for references to tasks
       for (const [service_name, service] of Object.entries(component.getServices())) {
         const node = graph.getNodeByRef(component.getServiceRef(service_name)) as ServiceNode;
         for (const interface_name of Object.keys(service.getInterfaces())) {
@@ -436,11 +482,14 @@ export default abstract class DependencyManager {
     const interpolated_environment_string = interpolateString(serialize(full_environment), context, ignore_keys);
     const environment = deserialize(this.environment.getClass(), interpolated_environment_string, { enableImplicitConversion: true }) as EnvironmentConfig;
 
-    // TODO:84:add support for references to tasks
     for (const component of Object.values(environment.getComponents())) {
       for (const [service_name, service] of Object.entries(component.getServices())) {
         const node = graph.getNodeByRef(component.getServiceRef(service_name)) as ServiceNode;
         node.node_config = service;
+      }
+      for (const [task_name, task] of Object.entries(component.getTasks())) {
+        const node = graph.getNodeByRef(component.getTaskRef(task_name)) as TaskNode;
+        node.node_config = task;
       }
     }
   }
